Manually save the view state of ViewPager2
The ViewPager2 implementation relies on a
RecyclerView to display its data, which is
fine, but it only restores the current item
position. When navigating back to a screen
containing a ViewPager2, the UI was not the
same as when the user left it, which I don't
think is a good UX.

Thus, by manually saving and restoring the
ViewState of the pager's pages, navigating
back to such a screen feels more natural.

This might lead to leaks since the new
adapter references view holders (thus a
context) that are not cleaned up.