Add a custom sqlite build
Since the search is an important feature of the app, it should be the best
possible. Using sqlite fts with Room makes the search very fast, but the
default tokenizer used to build the search index does not work well at
all with japanese text.

The same problem has been solved by Hai Feng Kao, introducing a custom
tokenizer: https://github.com/haifengkao/SqliteSubstringSearch
However, this means that we need to have a custom build of SqlLite to
be able to use it.

SQLite already has Android Bindings and instructions to build and
use a custom SQLite build in Android:
https://sqlite.org/android/doc/trunk/www/index.wiki
Jeremiah Stephenson also has a setup to build a custom SQLite version
for Android, and to provide custom tokenizers:
https://github.com/JeremiahStephenson/SQLite_Custom

I had to fork this repo to update it to a more recent version of SQLite
(3.31.1) and update the tooling, this is done in this fork:
https://github.com/esnaultdev/SQLite_Custom
I might upstream these modifications with a PR, but the repo has'nt been
updated for 3 years, so merging it might be slow.

Thankfully, Room provides a way to use a custom SQLite build, if you
bind the correct classes. This was pretty straightforward using the
framework sqlite binding as a template. However, I'm not sure why but
we now have to specify the DB path to properly create and open a db
now.

The search was updated to play a little with the tokenizer, which works
as expected. However, the search itself is a little bit broken, because
the match expressions need some tweaking.

Overall, this was tedious, but the result should pay off. I'm conflicted
about having to use a custom sqlite build, which will make the app
more bloated. But using LIKE '%term%' was not very scalable, and having
to migrate later to custom sqlite build would have been a pain to write
a migration for.