Fix the ghost review sync not working anymore
During the refactor to have a review primary key
based on both the review id and the review type,
the new DB mapping was not implemented correctly.
This resulted in foreign key violation errors
since the review history mapping was done properly
and referenced ghost reviews that were not in
the database.