Fix multiple textIsSelectable related issues
1) The text transition when toggling furigana
on and off was now making text without furigana
flash. This was due to textIsSelectable being
enabled on these views, making the TextView add
specific spans to the text. These spans were
making the text transition think that the text
changed, thus fading it out and in again.
The ChangeText transition was updated to
ignore those spans, but still take the ruby
spans into consideration.

2) Having the textIsSelectable flag enabled
on a TextView inside a RecyclerView item works
fine until the view gets recycled.
Updating the TextView to be wrap_content instead
of match_constraint fixes that.