Don't disable the edittext when answered
Disabling the edit text makes its colors change
a lot, and close the soft keyboard. We just
want the user to not be able to interact with
the edit text anymore.

isFocusable was already responsible for
preventing the edit text from receiving
the focus again. However, setting the
input type as null is the proper way to
prevent the user from pasting text or
continuing to use the keyboard.