Let the keyboard go to the next question
Having the input type set to null doesn't
close the keyboard, but we can't get input
from it.
Thus, let's keep the input type set to text
and disable more things about the edit text.

Typing still works though, so we need to
disable it using an edit text, but we need
to improve our wanakana binding to do that.