Update the default theme
Having the default theme of the app be the splash screen is only needed for the
spash screen itself. But it messes up with all previews in Android Studio since
that's the default theme of the app. Moving the splash screen theme to the
activity itself solved this problem by having the AppTheme be the default one.