Parse the review type of the current reviews.
Even though currently the API doesn't include the ghost reviews,
we can assume that it will do so in the future based on the field
review_type. This would not be a breaking change, so we need to
handle it properly, even though the ghost reviews are missing
right now.